refactor(module): 모듈 분리 및 어댑터에 뷰홀더 패턴 적용
