+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".confirm.ReservationConfirmActivity"
+            android:name=".ui.confirm.ReservationConfirmActivity"
             android:exported="false" />
         <activity
-            android:name=".reservation.MovieDetailActivity"
+            android:name=".ui.reservation.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".main.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
