--- app/src/main/java/woowacourse/movie/Toaster.kt
@@ -1,16 +1,0 @@
-package woowacourse.movie
-
-import android.content.Context
-import android.widget.Toast
-
-object Toaster {
-    private var toast: Toast? = null
-    fun showToast(context: Context, message: String) {
-        toast?.cancel()
-        toast = Toast.makeText(context, message, Toast.LENGTH_SHORT).also {
-            it.show()
-        }
-    }
-
-    fun cancel() = toast?.cancel()
-}
