--- app/src/main/java/woowacourse/movie/entity/Count.kt
@@ -1,16 +1,0 @@
-package woowacourse.movie.entity
-
-@JvmInline
-value class Count(val value: Int) : java.io.Serializable {
-    init {
-        require(value >= MIN_COUNT_RANGE) { COUNT_ERROR_MESSAGE }
-    }
-
-    operator fun minus(value: Int) = Count(this.value - value)
-    operator fun plus(value: Int) = Count(this.value + value)
-
-    companion object {
-        private const val COUNT_ERROR_MESSAGE = "[ERROR] 예약 인원 수는 1명 이상이여야 합니다."
-        private const val MIN_COUNT_RANGE = 1
-    }
-}
