--- app/src/test/java/woowacourse/movie/domain/DiscountCalculatorTest.kt
@@ -1,58 +1,0 @@
-package woowacourse.movie.domain
-
-class DiscountCalculatorTest {
-//    @Test
-//    fun `무비데이면 10% 할인을 받는다`() {
-//        val count = Count(2)
-//        val date = LocalDate.of(2023, 4, 10)
-//        val time = LocalTime.of(12, 0)
-//        val dateTime = LocalDateTime.of(date, time)
-//        val discountCalculator = DiscountCalculator()
-//        val actual = discountCalculator.discount(count, dateTime)
-//
-//        val expected = Money(23400)
-//        assertEquals(expected, actual)
-//    }
-//
-//    @Test
-//    fun `영화 시간대가 조조면 2000원 할인을 받는다`() {
-//        val count = Count(2)
-//        val date = LocalDate.of(2023, 4, 12)
-//        val time = LocalTime.of(10, 0)
-//        val dateTime = LocalDateTime.of(date, time)
-//        val discountCalculator = DiscountCalculator()
-//        val actual = discountCalculator.discount(count, dateTime)
-//
-//        val expected = Money(22000)
-//
-//        assertEquals(actual, expected)
-//    }
-//
-//    @Test
-//    fun `영화 시간대가 야간이면 2000원 할인을 받는다`() {
-//        val count = Count(3)
-//        val date = LocalDate.of(2023, 4, 12)
-//        val time = LocalTime.of(22, 0)
-//        val dateTime = LocalDateTime.of(date, time)
-//        val discountCalculator = DiscountCalculator()
-//        val actual = discountCalculator.discount(count, dateTime)
-//
-//        val expected = Money(33000)
-//
-//        assertEquals(actual, expected)
-//    }
-//
-//    @Test
-//    fun `무비데이이고 조조면 10%를 할인 받고 추가로 2000원 할인을 더 받는다`() {
-//        val count = Count(2)
-//        val date = LocalDate.of(2023, 4, 10)
-//        val time = LocalTime.of(10, 0)
-//        val dateTime = LocalDateTime.of(date, time)
-//        val discountCalculator = DiscountCalculator()
-//        val actual = discountCalculator.discount(count, dateTime)
-//
-//        val expected = Money(19400)
-//
-//        assertEquals(actual, expected)
-//    }
-}
