+++ domain/src/main/java/com/example/domain/model/Movie.kt
@@ -1,0 +1,12 @@
+package com.example.domain.model
+
+import java.time.LocalDate
+
+interface Movie {
+    val imgId: Int
+    val title: String
+    val startDate: LocalDate
+    val endDate: LocalDate
+    val runningTime: Int
+    val description: String
+}
