+++ domain/src/main/java/com/example/domain/setter/RunningDateSetter.kt
@@ -1,0 +1,19 @@
+package com.example.domain.setter
+
+import java.time.LocalDate
+import java.time.temporal.ChronoUnit.DAYS
+
+class RunningDateSetter {
+    fun getRunningDates(start: LocalDate, end: LocalDate): List<LocalDate> {
+        val betweenDays = DAYS.between(start, end)
+        val allDays = mutableListOf<LocalDate>()
+        for (i in ZERO..betweenDays) {
+            allDays.add(start.plusDays(i))
+        }
+        return allDays.toList()
+    }
+
+    companion object {
+        private const val ZERO = 0
+    }
+}
