+++ settings.gradle.kts
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "android-movie-ticket"
 include(":app")
+include(":domain")
